+++ docker-compose.yml
@@ -10,7 +10,7 @@ services:
       - POSTGRES_PASSWORD=123456
       - POSTGRES_DB=db_us_test
     ports:
-      - "5432:5432"
+      - "5433:5432"
 
   worker:
     build:
